rename config header
